Split out `ClientGenerics`
